I am a random traveler who seeks sensations like post-nut clarity.
Created a shared CSS and JS files. 
Basically, removed the redundant css and js files that keep repeating on each page. 
Added a function that automatically closes the overlay whenever the user clicks outside the box. 

Issues:
Compare overlay is not centered.
About overlay is not centered. 
I miss her so bad.